:zap: Optimize project start-up services
